+++ client/src/components/CreatePost/Form/Head.js
@@ -1,0 +1,15 @@
+import React from "react";
+import { Title, Badge, HeadWrapper } from "../StyledModal";
+
+const Head = ({ number, title }) => {
+  return (
+    <HeadWrapper>
+      <Badge>
+        <span>{number}</span>
+      </Badge>
+      <Title>{title}</Title>
+    </HeadWrapper>
+  );
+};
+
+export default Head;
